+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2024-03-08
+- Fix
+  - Closed orders not removed from order store upon PULL
+- Feature:
+  - option to get not completed orders for user
+
 ## [1.4.0] - 2024-02-21
 - Fix
   - WebSocket reconnect upon get token error (exponential backoff)